+++ .gitignore
@@ -1,0 +1,7 @@
+*~
+*#
+bin
+*.user*
+build
+*.sw?
+tags
