+++ glui/include/GL/glui.h
@@ -536,878 +536,6 @@ namespace glui {
      This is the only GLUI_Master_Object in existence.
   */
   extern GLUI_API GLUI_Master_Object GLUI_Master;
-
-  /************************************************************/
-  /*                                                          */
-  /*              Class for managing a GLUT window            */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     A top-level window.  The GLUI_Master GLUT callback can route events
-     to the callbacks in this class, for arbitrary use by external users.
-     (see GLUI_Master_Object::set_glutKeyboardFunc).
- 
-     This entire approach seems to be superceded by the "subwindow" flavor
-     of GLUI_Context *.
-  */
-  class GLUI_API GLUI_Glut_Window : public GLUI_Node 
-  {
-  public:
-    GLUI_Glut_Window();
-
-    int    glut_window_id;
-
-    /*********** Pointers to GLUT callthrough functions *****/
-    void (*glut_keyboard_CB)(unsigned char, int, int);
-    void (*glut_special_CB)(int, int, int);
-    void (*glut_reshape_CB)(int, int);
-    void (*glut_passive_motion_CB)(int,int);
-    void (*glut_mouse_CB)(int,int,int,int);
-    void (*glut_visibility_CB)(int);
-    void (*glut_motion_CB)(int,int);
-    void (*glut_display_CB)(void);
-    void (*glut_entry_CB)(int);
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*       Main Window GLUI_Context *class (not user-level)            */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     A GLUI_Main handles GLUT events for one window, routing them to the 
-     appropriate controls.  The central user-visible "GLUI_Context *" class 
-     inherits from this class; users should not allocate GLUT_Main objects.
-  
-     There's a separate GLUI_Main object for:
-     - Each top-level window with GUI stuff in it.
-     - Each "subwindow" of another top-level window.
-
-     All the GLUI_Main objects are listed in GLUI_Master.gluis.
-     A better name for this class might be "GLUI_Environment";
-     this class provides the window-level context for every control.
-  */
-  class GLUI_API GLUI_Main : public GLUI_Node 
-  {
-    /********** Friend classes *************/
-
-    friend class GLUI_Control;
-    friend class GLUI_Rotation;
-    friend class GLUI_Translation;
-    friend class GLUI_Context;
-    friend class GLUI_Master_Object;
-
-    /*********** Friend functions **********/
-
-    friend void glui_mouse_func(int button, int state, int x, int y);
-    friend void glui_keyboard_func(unsigned char key, int x, int y);
-    friend void glui_special_func(int key, int x, int y);
-    friend void glui_passive_motion_func(int x, int y);
-    friend void glui_reshape_func( int w, int h );
-    friend void glui_visibility_func(int state);
-    friend void glui_motion_func(int x, int y);
-    friend void glui_entry_func(int state);
-    friend void glui_display_func( void );
-    friend void glui_idle_func(void);
-
-    friend void glui_parent_window_reshape_func( int w, int h );
-    friend void glui_parent_window_keyboard_func( unsigned char, int, int );
-    friend void glui_parent_window_special_func( int, int, int );
-    friend void glui_parent_window_mouse_func( int, int, int, int );
-
-  protected:
-    /*** Variables ***/
-    int           main_gfx_window_id;
-    int           mouse_button_down;
-    int           glut_window_id;
-    int           top_level_glut_window_id;
-    GLUI_Control *active_control;
-    GLUI_Control *mouse_over_control;
-    GLUI_Panel   *main_panel;
-    enum buffer_mode_t {
-      buffer_front=1, ///< Draw updated controls directly to screen.
-      buffer_back=2   ///< Double buffering: postpone updates until next redraw.
-    };
-    buffer_mode_t buffer_mode; ///< Current drawing mode
-    int           curr_cursor;
-    int           w, h;
-    long          flags; 
-    bool          closing;
-    int           parent_window;
-    int           glui_id;
-
-    /********** Misc functions *************/
-
-    GLUI_Control  *find_control( int x, int y );
-    GLUI_Control  *find_next_control( GLUI_Control *control );
-    GLUI_Control  *find_next_control_rec( GLUI_Control *control );
-    GLUI_Control  *find_next_control_( GLUI_Control *control );
-    GLUI_Control  *find_prev_control( GLUI_Control *control );
-    void           create_standalone_window( const char *name, int x=-1, int y=-1 );
-    void           create_subwindow( int parent,int window_alignment );
-    void           setup_default_glut_callbacks( void );
-
-    void           mouse(int button, int state, int x, int y);
-    void           keyboard(unsigned char key, int x, int y);
-    void           special(int key, int x, int y);
-    void           passive_motion(int x, int y);
-    void           reshape( int w, int h );
-    void           visibility(int state);
-    void           motion(int x, int y);
-    void           entry(int state);
-    void           display( void );
-    void           idle(void);
-    int            needs_idle(void);
-
-    void (*glut_mouse_CB)(int, int, int, int);
-    void (*glut_keyboard_CB)(unsigned char, int, int);
-    void (*glut_special_CB)(int, int, int);
-    void (*glut_reshape_CB)(int, int);
-
-
-    /*********** Controls ************/
-
-    virtual int    add_control( GLUI_Node *parent, GLUI_Control *control );
-
-
-    /********** Constructors and Destructors ***********/
-
-    GLUI_Main( void );
-
-  public:
-    GLUI_StdBitmaps  std_bitmaps;
-    GLUI_String      window_name;
-    unsigned char    bkgd_color[3];
-    float            bkgd_color_f[3];
-
-    void            *font;
-    int              curr_modifiers;
-
-    void         adjust_glut_xy( int &x, int &y ) { y = h-y; }
-    void         activate_control( GLUI_Control *control, int how );
-    void         align_controls( GLUI_Control *control );
-    void         deactivate_current_control( void );
-    
-    /** Draw a 3D-look pushed-out box around this rectangle */
-    void         draw_raised_box( int x, int y, int w, int h );
-    /** Draw a 3D-look pushed-in box around this rectangle */
-    void         draw_lowered_box( int x, int y, int w, int h );
-    
-    /** Return true if this control should redraw itself immediately (front buffer);
-        Or queue up a redraw and return false if it shouldn't (back buffer).
-    */
-    bool         should_redraw_now(GLUI_Control *ctl);
-    
-    /** Switch to the appropriate draw buffer now.  Returns the old draw buffer. 
-        This routine should probably only be called from inside the GLUI_DrawingSentinal,
-        in glui_internal_control.h
-    */
-    int          set_current_draw_buffer();
-    /** Go back to using this draw buffer.  Undoes set_current_draw_buffer. */
-    void         restore_draw_buffer( int buffer_state );
-    
-    /** Pack, resize the window, and redraw all the controls. */
-    void         refresh();
-    
-    /** Redraw the main graphics window */
-    void         post_update_main_gfx();
-  
-    /** Recompute the sizes and positions of all controls */
-    void         pack_controls();
-    
-    void         close_internal();
-    void         check_subwindow_position();
-    void         set_ortho_projection();
-    void         set_viewport();
-    int          get_glut_window_id( void ) { return glut_window_id; } /* JVK */
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*       GLUI_Control: base class for all controls          */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     All the GUI objects inherit from GLUI_Control: buttons,
-     checkboxes, labels, edit boxes, scrollbars, etc.
-     Most of the work of this class is in routing events,
-     like keystrokes, mouseclicks, redraws, and sizing events.
- 
-     Yes, this is a huge and hideous class.  It needs to be 
-     split up into simpler subobjects.  None of the data members
-     should be directly accessed by users (they should be protected,
-     not public); only subclasses.
-  */
-  class GLUI_API GLUI_Control : public GLUI_Node 
-  {
-  public:
-
-    /** Onscreen coordinates */
-    int             w, h;                        /* dimensions of control */
-    int             x_abs, y_abs;
-    int             x_off, y_off_top, y_off_bot; /* INNER margins, by which child
-                                                    controls are indented */
-    int             contain_x, contain_y; 
-    int             contain_w, contain_h;
-    /* if this is a container control (e.g., 
-       radiogroup or panel) this indicated dimensions
-       of inner area in which controls reside */
-
-    /** "activation" for tabbing between controls. */
-    int             active_type; ///< "GLUI_CONTROL_ACTIVE_..."
-    bool            active;       ///< If true, we've got the focus
-    bool            can_activate; ///< If false, remove from tab order.
-    bool            spacebar_mouse_click; ///< Spacebar simulates click.
-    
-    /** Callbacks */
-    long            user_id;  ///< Integer to pass to callback function.
-    CallBack         callback; ///< User callback function, or NULL.
-
-    /** Variable value storage */
-    float           float_val;        /**< Our float value */
-    int             int_val;          /**< Our integer value */
-    float           float_array_val[GLUI_DEF_MAX_ARRAY];
-    int             float_array_size;
-    GLUI_String     text;       /**< The text inside this control */
-    
-    /** "Live variable" updating */
-    void           *ptr_val { nullptr };          /**< A pointer to the user's live variable value */
-    int             live_type;
-    bool            live_inited;
-    /* These variables store the last value that live variable was known to have. */
-    int             last_live_int;  
-    float           last_live_float;
-    GLUI_String     last_live_text;
-    float           last_live_float_array[GLUI_DEF_MAX_ARRAY];
-    
-    /** Properties of our control */    
-    GLUI_Context           *context { nullptr };       /**< Our containing event handler (NEVER NULL during event processing!) */
-    bool            is_container;  /**< Is this a container class (e.g., panel) */
-    int             alignment;
-    bool            enabled;    /**< Is this control grayed out? */
-    GLUI_String     name;       /**< The name of this control */
-    void           *font;       /**< Our glutbitmap font */
-    bool            collapsible, is_open;
-    GLUI_Node       collapsed_node;
-    bool            hidden; /* Collapsed controls (and children) are hidden */
-    int             char_widths[CHAR_WIDTH_HASH_SIZE][2]; /* Character width hash table */
-
-  public:
-    /*** Get/Set values ***/
-    virtual void   set_name( const char *string );
-    virtual void   set_int_val( int new_int )         { int_val = new_int; output_live(true); }
-    virtual void   set_float_val( float new_float )   { float_val = new_float; output_live(true); }
-    virtual void   set_ptr_val( void *new_ptr )       { ptr_val = new_ptr; output_live(true); }
-    virtual void   set_float_array_val( float *array_ptr );
-
-    virtual float  get_float_val( void )              { return float_val; }
-    virtual int    get_int_val( void )                { return int_val; }
-    virtual void   get_float_array_val( float *array_ptr );
-    virtual int    get_id( void ) const { return user_id; }
-    virtual void   set_id( int id ) { user_id=id; }
-
-    virtual int mouse_down_handler( int local_x, int local_y )                 { return false; }
-    virtual int mouse_up_handler( int local_x, int local_y, bool inside )       { return false; }
-    virtual int mouse_held_down_handler( int local_x, int local_y, bool inside) { return false; }
-    virtual int key_handler( unsigned char key, int modifiers )                { return false; }
-    virtual int special_handler( int key,int modifiers )                       { return false; }
-
-    virtual void update_size( void )     { }
-    virtual void idle( void )            { }
-    virtual int  mouse_over( int state, int x, int y ) { return false; }
-
-    virtual void enable( void ); 
-    virtual void disable( void );
-    virtual void activate( int how )     { active = true; }
-    virtual void deactivate( void )     { active = false; }
-
-    /** Hide (shrink into a rollout) and unhide (expose from a rollout) */
-    void         hide_internal( int recurse );
-    void         unhide_internal( int recurse );
-
-    /** Return true if it currently makes sense to draw this class. */
-    int          can_draw( void ) { return (context != nullptr && hidden == false); }
-
-    /** Redraw this control.
-        In single-buffering mode (drawing to GL_FRONT), this is just 
-        a call to translate_and_draw_front (after a can_draw() check).
-        In double-buffering mode (drawing to GL_BACK), this queues up 
-        a redraw and returns false, since you shouldn't draw yet.
-    */
-    void          redraw(void);
-    
-    /** Redraw everybody in our window. */
-    void         redraw_window(void);
-
-    virtual void align( void );
-    void         pack( int x, int y );    /* Recalculate positions and offsets */
-    void         pack_old( int x, int y );    
-    void         draw_recursive( int x, int y );
-    int          set_to_glut_window( void );
-    void         restore_window( int orig );
-    void         translate_and_draw_front( void );
-    void         translate_to_origin( void ) 
-    {glTranslatef((float)x_abs+.5,(float)y_abs+.5,0.0);}
-    virtual void draw( int x, int y )=0;
-    void         set_font( void *new_font );
-    void        *get_font( void );
-    int          string_width( const char *text );
-    int          string_width( const GLUI_String &str ) 
-    { return string_width(str.c_str()); }
-    int          char_width( char c );
-
-    void         draw_name( int x, int y );
-    void         draw_box_inwards_outline( int x_min, int x_max, 
-                                           int y_min, int y_max );
-    void         draw_box( int x_min, int x_max, int y_min, int y_max,
-                           float r, float g, float b );
-    void         draw_bkgd_box( int x_min, int x_max, int y_min, int y_max );
-    void         draw_emboss_box( int x_min, int x_max,int y_min,int y_max);
-    void         draw_string( const char *text );
-    void         draw_string( const GLUI_String &s ) 
-    { draw_string(s.c_str()); }
-    void         draw_char( char c );
-    void         draw_active_box( int x_min, int x_max, int y_min, int y_max );
-    void         set_to_bkgd_color( void );
-
-    void         set_w( int new_w );
-    void         set_h( int new_w );
-    void         set_alignment( int new_align );
-    void         sync_live( int recurse, int draw );  /* Reads live variable */
-    void         init_live( void );
-    void         output_live( int update_main_gfx );        /** Writes live variable **/
-    virtual void set_text( const char *t )   {}
-    void         execute_callback( void );
-    void         get_this_column_dims( int *col_x, int *col_y, 
-                                       int *col_w, int *col_h, 
-                                       int *col_x_off, int *col_y_off );
-    virtual bool needs_idle( void ) const;
-    virtual bool wants_tabs() const      { return false; }
-
-    GLUI_Control(void) 
-      {
-        x_off          = GLUI_XOFF;
-        y_off_top      = GLUI_YOFF;
-        y_off_bot      = GLUI_YOFF;
-        x_abs          = GLUI_XOFF;
-        y_abs          = GLUI_YOFF;
-        active         = false;
-        enabled        = true;
-        int_val        = 0;
-        last_live_int  = 0;
-        float_array_size = 0;
-        glui_format_str(name, "Control: %p", this);
-        float_val      = 0.0;
-        last_live_float = 0.0;
-        w              = GLUI_DEFAULT_CONTROL_WIDTH;
-        h              = GLUI_DEFAULT_CONTROL_HEIGHT;
-        font           = NULL;
-        active_type    = GLUI_CONTROL_ACTIVE_MOUSEDOWN;
-        alignment      = GLUI_ALIGN_LEFT;
-        is_container   = false;
-        can_activate   = true;         /* By default, you can activate a control */
-        spacebar_mouse_click = true;    /* Does spacebar simulate a mouse click? */
-        live_type      = GLUI_LIVE_NONE;
-        text = "";
-        last_live_text = "";
-        live_inited    = false;
-        collapsible    = false;
-        is_open        = true;
-        hidden         = false;
-        memset(char_widths, -1, sizeof(char_widths)); /* JVK */
-        int i;
-        for( i=0; i<GLUI_DEF_MAX_ARRAY; i++ )
-          float_array_val[i] = last_live_float_array[i] = 0.0;
-      }
-
-    virtual ~GLUI_Control();
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               Button class (container)                   */
-  /*                                                          */
-  /************************************************************/
-  /**
-     An onscreen, clickable button--an outlined label that 
-     can be clicked.  When clicked, a button
-     calls its CallBack callback with its ID.
-  */
-  class GLUI_API GLUI_Button : public GLUI_Control
-  {
-  public:
-    bool currently_inside;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-
-    void draw( int x, int y );
-    void draw_pressed( void );
-    void draw_text( int sunken );
-
-    void update_size( void );
-
-    /**
-       Create a new button.
- 
-       @param parent The panel our object is inside; or the main GLUI_Context *object.
-       @param name The text inside the button.
-       @param id Optional ID number, to pass to the optional callback function.
-       @param callback Optional callback function, taking either the int ID or control.
-    */
-    GLUI_Button( GLUI_Node *parent, const char *name, 
-                 int id=-1, CallBack cb=CallBack() );
-    GLUI_Button( void ) { common_init(); };
-
-  protected:
-    void common_init(void) {
-      glui_format_str(name, "Button: %p", this );
-      h            = GLUI_BUTTON_SIZE;
-      w            = 100;
-      alignment    = GLUI_ALIGN_CENTER;
-      can_activate = true;
-    }
-  };
-
-
-  /************************************************************/
-  /*                                                          */
-  /*               Checkbox class (container)                 */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     A checkbox, which can be checked on or off.  Can be linked
-     to an int value, which gets 1 for on and 0 for off.
-  */
-  class GLUI_API GLUI_Checkbox : public GLUI_Control
-  {
-  public:
-    int  orig_value;
-    bool currently_inside;
-    int  text_x_offset;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-
-    void update_size( void );
-
-    void draw( int x, int y );
-
-    void draw_active_area( void );
-    void draw_empty_box( void );
-    void set_int_val( int new_val );
-
-    /**
-       Create a new checkbox object.
- 
-       @param parent The panel our object is inside; or the main GLUI_Context *object.
-       @param name Label next to our checkbox.
-       @param value_ptr Optional integer value to attach to this checkbox.  When the 
-       checkbox is checked or unchecked, *value_ptr will also be changed. ("Live Vars").
-       @param id Optional ID number, to pass to the optional callback function.
-       @param callback Optional callback function, taking either the int ID or control.
-    */
-    GLUI_Checkbox(GLUI_Node *parent, const char *name, int *value_ptr=NULL,
-                  int id=-1, CallBack callback=CallBack());
-    GLUI_Checkbox( void ) { common_init(); }
-
-  protected:
-    void common_init(void) {
-      glui_format_str( name, "Checkbox: %p", this );
-      w              = 100;
-      h              = GLUI_CHECKBOX_SIZE;
-      orig_value     = -1;
-      text_x_offset  = 18;
-      can_activate   = true;
-      live_type      = GLUI_LIVE_INT;   /* This control has an 'int' live var */
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               Column class                               */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     A GLUI_Column object separates all previous controls
-     from subsequent controls with a vertical bar.
-  */
-  class GLUI_API GLUI_Column : public GLUI_Control
-  {
-  public:
-    void draw( int x, int y );
-
-    /**
-       Create a new column, which separates the previous controls
-       from subsequent controls.
- 
-       @param parent The panel our object is inside; or the main GLUI_Context *object.
-       @param draw_bar If true, draw a visible bar between new and old controls.
-    */
-    GLUI_Column( GLUI_Node *parent, int draw_bar = true );
-    GLUI_Column( void ) { common_init(); }
-
-  protected:
-    void common_init() {
-      w            = 0;
-      h            = 0;
-      int_val      = 0;
-      can_activate = false;
-    }
-  };
-
-
-  /************************************************************/
-  /*                                                          */
-  /*               Panel class (container)                    */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     A GLUI_Panel contains a group of related controls.
-  */
-  class GLUI_API GLUI_Panel : public GLUI_Control
-  {
-  public:
-
-    /**
-       Create a new panel.  A panel groups together a set of related controls.
- 
-       @param parent The outer panel our panel is inside; or the main GLUI_Context *object.
-       @param name The string name at the top of our panel.
-       @param type Optional style to display the panel with--GLUI_PANEL_EMBOSSED by default.
-       GLUI_PANEL_RAISED causes the panel to appear higher than the surroundings.
-       GLUI_PANEL_NONE causes the panel's outline to be invisible.
-    */
-    GLUI_Panel( GLUI_Node *parent, const char *name, 
-                int type=GLUI_PANEL_EMBOSSED );
-    GLUI_Panel() { common_init(); }
-
-    void draw( int x, int y );
-    void set_name( const char *text );
-    void set_type( int new_type );
-
-    void update_size( void );
-
-  protected:
-    void common_init( void ) {
-      w            = 300;
-      h            = GLUI_DEFAULT_CONTROL_HEIGHT + 7;
-      int_val      = GLUI_PANEL_EMBOSSED;
-      alignment    = GLUI_ALIGN_CENTER;
-      is_container = true; 
-      can_activate = false;
-      name="";
-    };
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               File Browser class (container)             */
-  /*                         JVK                              */
-  /************************************************************/
-
-  /**
-     A list of files the user can select from.
-  */
-  class GLUI_API GLUI_FileBrowser : public GLUI_Panel
-  {
-  public:
-    /**
-       Create a new list of files the user can select from.
- 
-       @param parent The panel our object is inside; or the main GLUI_Context *object.
-       @param name Prompt to give to the user at the top of the file browser.
-       @param frame_type Optional style to display the panel with--GLUI_PANEL_EMBOSSED by default.
-       GLUI_PANEL_RAISED causes the panel to appear higher than the surroundings.
-       GLUI_PANEL_NONE causes the panel's outline to be invisible.
-       @param id Optional ID number, to pass to the optional callback function.
-       @param callback Optional callback function, taking either the int ID or control.
-    */
-    GLUI_FileBrowser( GLUI_Node *parent, 
-                      const char *name,
-                      int frame_type = GLUI_PANEL_EMBOSSED,
-                      int user_id = -1,
-                      CallBack callback = CallBack());
-
-    GLUI_List *list;
-    GLUI_String current_dir;
-
-    void fbreaddir(const char *);
-    static void dir_list_callback(GLUI_Control*);
-
-    void set_w(int w);
-    void set_h(int h);
-    const char* get_file() { return file.c_str(); }
-    void set_allow_change_dir(int c) { allow_change_dir = c; }
-
-  protected:
-    void common_init() 
-    {
-      w            = GLUI_DEFAULT_CONTROL_WIDTH;
-      h            = GLUI_DEFAULT_CONTROL_HEIGHT;
-      int_val      = GLUI_PANEL_EMBOSSED;
-      alignment    = GLUI_ALIGN_CENTER;
-      is_container = true; 
-      can_activate = false;
-      allow_change_dir = true;
-      last_item    = -1;
-      user_id      = -1;
-      name         = "";
-      current_dir  = ".";
-      file         = "";
-    };
-
-  private:
-    int last_item;
-    GLUI_String file;
-    int allow_change_dir;
-
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               Rollout class (container)                  */
-  /*                                                          */
-  /************************************************************/
-  /**
-     A rollout contains a set of controls,
-     like a panel, but can be collapsed to just the name.
-  */
-  class GLUI_API GLUI_Rollout : public GLUI_Panel
-  {
-  public:
-
-    /**
-       Create a new rollout.  A rollout contains a set of controls,
-       like a panel, but can be collapsed to just the name.
- 
-       @param parent The panel our object is inside; or the main GLUI_Context *object.
-       @param name String to show at the top of the rollout.
-       @param open Optional boolean.  If true (the default), the rollout's controls are displayed.
-       If false, the rollout is closed to display only the name.
-       @param type Optional style to display the panel with--GLUI_PANEL_EMBOSSED by default.
-       GLUI_PANEL_RAISED causes the panel to appear higher than the surroundings.
-       GLUI_PANEL_NONE causes the panel's outline to be invisible.
-    */
-    GLUI_Rollout( GLUI_Node *parent, const char *name, int open=true, 
-                  int type=GLUI_PANEL_EMBOSSED );
-    GLUI_Rollout( void ) { common_init(); }
-    
-    
-    bool        currently_inside, initially_inside;
-    GLUI_Button  button;
-
-    void draw( int x, int y );
-    void draw_pressed( void );
-    int mouse_down_handler( int local_x, int local_y );
-    int mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-        
-    void  open( void ); 
-    void  close( void );
-
-    void update_size( void );
-
-  protected:
-    void common_init() {
-      currently_inside = false;
-      initially_inside = false;
-      can_activate     = true;
-      is_container     = true;
-      h                = GLUI_DEFAULT_CONTROL_HEIGHT + 7;
-      w                = GLUI_DEFAULT_CONTROL_WIDTH;
-      y_off_top        = 21;
-      collapsible      = true;
-      name = "";
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               Tree    Panel class (container)            */
-  /*                         JVK                              */
-  /************************************************************/
-
-  /**
-     One collapsible entry in a GLUI_TreePanel.
-  */
-  class GLUI_API GLUI_Tree : public GLUI_Panel
-  {
-  public:
-    GLUI_Tree(GLUI_Node *parent, const char *name, 
-              int open=false, int inset=0);
-
-  private:
-    int level;   // how deep is this node
-    float red;   //Color coding of column line
-    float green;
-    float blue;
-    float lred;   //Color coding of level name
-    float lgreen;
-    float lblue;
-    int id;
-    GLUI_Column *column;
-    int is_current;          // Whether this tree is the
-    // current root in a treePanel
-    int child_number;
-    int format;
-
-  public:
-    bool        currently_inside, initially_inside;
-    GLUI_Button  button;
-    GLUI_String  level_name; // level name, eg: 1.1.2, III, or 3
-    GLUI_TreePanel *panel; 
-
-    void draw( int x, int y );
-    void draw_pressed( void );
-    int mouse_down_handler( int local_x, int local_y );
-    int mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    void set_column(GLUI_Column *c) { column = c; }
-    void  open( void ); 
-    void  close( void );
-
-    /*   void set_name( const char *text )   { panel.set_name( text ); }; */
-    void update_size( void );
-    void set_id(int i) { id = i; }
-    void set_level(int l) { level = l; }
-    void set_format(int f) { format = f; }
-    void set_current(int c) { is_current = c; }
-    int get_id() { return id; }
-    int get_level() { return level; }
-    int get_child_number() { return child_number; }
-    void enable_bar() { if (column) { column->int_val = 1;  set_color(red, green, blue); } }
-    void disable_bar() { if (column) { column->int_val = 0;  } } 
-    void set_child_number(int c) { child_number = c; } 
-    void set_level_color(float r, float g, float b) { 
-      lred = r;
-      lgreen = g;
-      lblue  = b;
-    }
-    void set_color(float r, float g, float b) { 
-      red = r;
-      green = g;
-      blue  = b;
-    }
-  protected:
-    void common_init()
-    {
-      currently_inside = false;
-      initially_inside = false;
-      can_activate     = true;
-      is_container     = true;
-      h                = GLUI_DEFAULT_CONTROL_HEIGHT + 7;
-      w                = GLUI_DEFAULT_CONTROL_WIDTH;
-      y_off_top        = 21;
-      collapsible      = true;
-      red              = .5;
-      green            = .5;
-      blue             = .5;
-      lred             = 0;
-      lgreen           = 0;
-      lblue            = 0;
-      column           = NULL;
-      is_current       = 0;
-      child_number     = 0;
-      format           = 0;
-      panel            = NULL;
-      name             = "";
-      level_name       = "";
-      level            = 0;
-    
-    };
-  };
-
-
-  /************************************************************/
-  /*                                                          */
-  /*               TreePanel class (container) JVK            */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     Manages, maintains, and formats a tree of GLUI_Tree objects.
-     These are shown in a heirarchical, collapsible display.
-  
-     FIXME: There's an infinite loop in the traversal code (OSL 2006/06)
-  */
-  class GLUI_API GLUI_TreePanel : public GLUI_Panel 
-  {
-  public:
-    GLUI_TreePanel(GLUI_Node *parent, const char *name,
-                   bool open=false, int inset=0);
-
-    int max_levels;
-    int next_id;
-    int format;
-    float red;
-    float green;
-    float blue;
-    float lred;
-    float lgreen;
-    float lblue;
-    int root_children;
-    /* These variables allow the tree panel to traverse the tree
-       using only two function calls. (Well, four, if you count 
-       going in reverse */
-
-    GLUI_Tree    *curr_branch; /* Current Branch */
-    GLUI_Panel *curr_root;   /* Current Root */
-
-  public:
-    void            set_color(float r, float g, float b); 
-    void            set_level_color(float r, float g, float b);
-    void            set_format(int f) { format = f; }
-
-    /* Adds branch to curr_root */
-    GLUI_Tree *     ab(const char *name, GLUI_Tree *root = NULL);
-    /* Goes up one level, resets curr_root and curr_branch to parents*/
-    void            fb(GLUI_Tree *branch= NULL);
-    /* Deletes the curr_branch, goes up one level using fb */
-    void            db(GLUI_Tree *branch = NULL);
-    /* Finds the very last branch of curr_root, resets vars */
-    void            descendBranch(GLUI_Panel *root = NULL);
-    /* Resets curr_root and curr branch to TreePanel and lastChild */
-    void            resetToRoot(GLUI_Panel *new_root = NULL);
-    void            next( void );
-    void            refresh( void );
-    void            expand_all( void );
-    void            collapse_all( void );
-    void            update_all( void );
-    void            initNode(GLUI_Tree *temp);
-    void            formatNode(GLUI_Tree *temp);
-
-  protected:
-    int uniqueID( void ) { next_id++; return next_id - 1; }
-    void common_init() 
-    {
-      GLUI_Panel();
-      next_id = 0;
-      curr_root = this;
-      curr_branch = NULL;
-      red = .5;
-      green = .5;
-      blue = .5;
-      root_children = 0;
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*                     User-Level GLUI_Context *class                */
-  /*                                                          */
-  /************************************************************/
-
   class GLUI_Rotation;
   class GLUI_Translation;
 
@@ -1426,6 +554,17 @@ namespace glui {
 }
 
 
+#include "glui/GlutWindow.h"
+#include "glui/Main.h"
+#include "glui/Control.h"
+#include "glui/Panel.h"
+#include "glui/Button.h"
+#include "glui/Column.h"
+#include "glui/CheckBox.h"
+#include "glui/FileBrowser.h"
+#include "glui/RollOut.h"
+#include "glui/Tree.h"
+#include "glui/TreePanel.h"
 #include "glui/Context.h"
 #include "glui/Spinner.h"
 #include "glui/Separator.h"
